+++ dmr/security/jwt/__init__.py
@@ -11,6 +11,10 @@ except ImportError:  # pragma: no cover
     )
     raise
 
+from dmr.security.jwt.auth import BaseJWTAsyncAuth as BaseJWTAsyncAuth
+from dmr.security.jwt.auth import BaseJWTSyncAuth as BaseJWTSyncAuth
+from dmr.security.jwt.auth import HeaderJWTAsyncAuth as HeaderJWTAsyncAuth
+from dmr.security.jwt.auth import HeaderJWTSyncAuth as HeaderJWTSyncAuth
 from dmr.security.jwt.auth import JWTAsyncAuth as JWTAsyncAuth
 from dmr.security.jwt.auth import JWTSyncAuth as JWTSyncAuth
 from dmr.security.jwt.auth import request_jwt as request_jwt